feat: 쿠폰 도메인

+++ src/main/kotlin/com/ps/app/coupons/adapter/in/web/dto/IssueCouponRequest.kt
@@ -5,7 +5,7 @@ package com.ps.app.coupons.adapter.`in`.web.dto
  */
 data class IssueCouponRequest(
     val ownerId: Long,
-    val couponPolicyId: Long,
+    val couponPolicyId: Int,
     val couponCode: String,
     val validDays: Int
 )
