+++ src/main/kotlin/com/ps/app/coupons/application/port/in/IssueCouponCommand.kt
@@ -2,7 +2,7 @@ package com.ps.app.coupons.application.port.`in`
 
 data class IssueCouponCommand(
     val ownerId: Long,
-    val couponPolicyId: Long,
+    val couponPolicyId: Int,
     val couponCode: String,
     val validDays: Int
 )
